# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UsrResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"lastClickAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"lastName",
					"email",
					"lastClickAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUsrDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"lastClickAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"CreateUsrDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"lastName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Knight API",
		"version": "1.0.0",
		"description": "API para gestión de caballeros - Documentación automática con TypeScript"
	},
	"paths": {
...
		"/api/usrs": {
			"get": {
				"operationId": "Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UsrResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Usr"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "term",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUsr",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsrResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Usr"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUsrDTO"
							}
						}
					}
				}
			}
		},
		"/api/usrs/{id}": {
			"put": {
				"operationId": "UpdateUsr",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsrResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Usr"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUsrDTO"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}